+++ Sources/MapboxMaps/Foundation/Extensions/Core/ResourceOptions.swift
@@ -15,13 +15,10 @@ public struct ResourceOptions {
 
     /// The file URL to the cache. The default, `nil`, will choose an appropriate
     /// location on the device. The default location is excluded from backups.
-    public var cachePathURL: URL?
+    public var dataPathURL: URL?
 
     /// The path to the assets. The default, `nil`, uses the main bundle.
     public var assetPathURL: URL?
-
-    /// The size of the cache in bytes. The default, `nil`, uses `defaultCacheSize`
-    public var cacheSize: UInt64?
 
     /// The tile store instance
     ///
@@ -34,7 +31,7 @@ public struct ResourceOptions {
     /// - Attention:
     ///     If you create a `ResourceOptions` (rather than using `ResourceOptionsManager`
     ///     to manage one) that uses a custom TileStore, you will need to ensure
-    ///     that the `TileStore` is initialised with a valid access token.
+    ///     that the `TileStore` is initialized with a valid access token.
     ///
     ///     For example:
     ///
@@ -46,18 +43,16 @@ public struct ResourceOptions {
     /// Tile store usage mode
     public var tileStoreUsageMode: TileStoreUsageMode
 
-    /// The default size of the cache. Used if `cacheSize` is set to `nil`
-    public static let defaultCacheSize: UInt64 = (1024*1024*50)
-
     /// Initialize a `ResourceOptions`, used by both `MapView`s and `Snapshotter`s
     /// - Parameters:
     ///   - accessToken: Mapbox access token. You must provide a valid token.
     ///   - baseUrl: Base url for resource requests; default is `nil`
-    ///   - cachePath: Path to database cache; default is `nil`, which will create
-    ///         a path in the application's "support directory"
-    ///   - assetPath: Path to assets; default is `nil`, which will use the application's
-    ///         resource bundle. `assetPath` is expected to be path to a bundle.
-    ///   - cacheSize: Size of the cache.
+    ///   - dataPathURL: Path to database cache; default is `nil`, which will create
+    ///         a path in a suitable application directory that is excluded from
+    ///         backups.
+    ///   - assetPathURL: Path to assets; default is `nil`, which will use the
+    ///         application's resource bundle. `assetPath` is expected to be path
+    ///         to a bundle.
     ///   - tileStore: A tile store is only used if `tileStoreEnabled` is `true`,
     ///         otherwise this argument is ignored. If `nil` (and `tileStoreEnabled`
     ///         is `true`, the a default tile store will be created and used.
@@ -88,40 +83,16 @@ public struct ResourceOptions {
     ///     exists in the cache.
     public init(accessToken: String,
                 baseURL: URL? = nil,
-                cachePathURL: URL? = nil,
+                dataPathURL: URL? = nil,
                 assetPathURL: URL? = nil,
-                cacheSize: UInt64 = Self.defaultCacheSize,
                 tileStore: TileStore? = nil,
                 tileStoreUsageMode: TileStoreUsageMode = .readOnly) {
         self.accessToken        = accessToken
         self.baseURL            = baseURL
-        self.cachePathURL       = cachePathURL ?? ResourceOptions.cacheURLIncludingSubdirectory()
+        self.dataPathURL        = dataPathURL
         self.assetPathURL       = assetPathURL ?? Bundle.main.resourceURL
-        self.cacheSize          = cacheSize
         self.tileStore          = tileStore
         self.tileStoreUsageMode = tileStoreUsageMode
-    }
-
-    private static func cacheURLIncludingSubdirectory() -> URL {
-        do {
-            var cacheDirectoryURL = try FileManager.default.url(for: .applicationSupportDirectory,
-                                                                in: .userDomainMask,
-                                                                appropriateFor: nil,
-                                                                create: true)
-
-            cacheDirectoryURL = cacheDirectoryURL.appendingPathComponent(".mapbox")
-            cacheDirectoryURL = cacheDirectoryURL.appendingPathComponent("maps")
-
-            try FileManager.default.createDirectory(at: cacheDirectoryURL,
-                                                    withIntermediateDirectories: true,
-                                                    attributes: nil)
-
-            cacheDirectoryURL.setTemporaryResourceValue(true, forKey: .isExcludedFromBackupKey)
-
-            return cacheDirectoryURL.appendingPathComponent("ambient_cache.db")
-        } catch {
-            fatalError("Failed to create cache directory: \(error)")
-        }
     }
 }
 
@@ -132,9 +103,10 @@ extension ResourceOptions: Hashable {
             && ((lhs.baseURL == rhs.baseURL)
                     || (lhs.baseURL == nil)
                     || (rhs.baseURL == nil))
-            && (lhs.cachePathURL == rhs.cachePathURL)
+            && ((lhs.dataPathURL == rhs.dataPathURL)
+                || (lhs.dataPathURL == nil)
+                || (rhs.dataPathURL == nil))
             && (lhs.assetPathURL == rhs.assetPathURL)
-            && (lhs.cacheSize == rhs.cacheSize)
             && ((lhs.tileStore == rhs.tileStore)
                     || (lhs.tileStore == nil)
                     || (rhs.tileStore == nil))
@@ -145,9 +117,8 @@ extension ResourceOptions: Hashable {
     public func hash(into hasher: inout Hasher) {
         hasher.combine(accessToken)
         hasher.combine(baseURL)
-        hasher.combine(cachePathURL)
+        hasher.combine(dataPathURL)
         hasher.combine(assetPathURL)
-        hasher.combine(cacheSize)
         hasher.combine(tileStore)
         hasher.combine(tileStoreUsageMode)
     }
@@ -181,14 +152,13 @@ extension ResourceOptions {
     internal init(_ objcValue: MapboxCoreMaps.ResourceOptions) {
 
         let baseURL      = objcValue.baseURL.flatMap { URL(fileURLWithPath: $0) }
-        let cachePathURL = objcValue.cachePath.flatMap { URL(fileURLWithPath: $0) }
+        let dataPathURL = objcValue.dataPath.flatMap { URL(fileURLWithPath: $0) }
         let assetPathURL = objcValue.assetPath.flatMap { URL(fileURLWithPath: $0) }
 
         self.init(accessToken: objcValue.accessToken,
                   baseURL: baseURL,
-                  cachePathURL: cachePathURL,
+                  dataPathURL: dataPathURL,
                   assetPathURL: assetPathURL,
-                  cacheSize: objcValue.__cacheSize?.uint64Value ?? Self.defaultCacheSize,
                   tileStore: objcValue.tileStore,
                   tileStoreUsageMode: objcValue.tileStoreUsageMode)
     }
@@ -198,9 +168,8 @@ extension MapboxCoreMaps.ResourceOptions {
     internal convenience init(_ swiftValue: ResourceOptions) {
         self.init(__accessToken: swiftValue.accessToken,
                   baseURL: swiftValue.baseURL?.path,
-                  cachePath: swiftValue.cachePathURL?.path,
+                  dataPath: swiftValue.dataPathURL?.path,
                   assetPath: swiftValue.assetPathURL?.path,
-                  cacheSize: swiftValue.cacheSize?.NSNumber,
                   tileStore: swiftValue.tileStore,
                   tileStoreUsageMode: swiftValue.tileStoreUsageMode)
     }