+++ Sources/MapboxMaps/Location/Pucks/Puck.swift
@@ -1,12 +1,11 @@
-#if canImport(MapboxMapsFoundation)
-import MapboxMapsFoundation
-#endif
-
 /// This protocol is used to help manipulate the different type of puck views we have
 internal protocol Puck {
 
     /// Property that stores the current `PuckStyle` of the puck
     var puckStyle: PuckStyle { get set }
+
+    /// Property that stores the current `PuckBearingSource` of the puck
+    var puckBearingSource: PuckBearingSource { get set }
 
     /// Property that references the mapView that the puck should be draw
     var locationSupportableMapView: LocationSupportableMapView? { get }