+++ Sources/MapboxMaps/Location/Pucks/Puck2D.swift
@@ -3,14 +3,6 @@ import MapboxCoreMaps
 @_implementationOnly import MapboxCoreMaps_Private
 @_implementationOnly import MapboxCommon_Private
 
-#if canImport(MapboxMapsFoundation)
-import MapboxMapsFoundation
-#endif
-
-#if canImport(MapboxMapsStyle)
-import MapboxMapsStyle
-#endif
-
 public struct Puck2DConfiguration: Equatable {
 
     /// Image to use as the top of the location indicator.
@@ -56,16 +48,19 @@ internal class Puck2D: Puck {
 
     // MARK: Protocol Properties
     internal var puckStyle: PuckStyle
+    internal var puckBearingSource: PuckBearingSource
 
     internal weak var locationSupportableMapView: LocationSupportableMapView?
     internal weak var style: LocationStyleDelegate?
 
     // MARK: Initializers
     internal init(puckStyle: PuckStyle,
+                  puckBearingSource: PuckBearingSource,
                   locationSupportableMapView: LocationSupportableMapView,
                   style: LocationStyleDelegate,
                   configuration: Puck2DConfiguration) {
         self.puckStyle = puckStyle
+        self.puckBearingSource = puckBearingSource
         self.locationSupportableMapView = locationSupportableMapView
         self.style = style
         self.configuration = configuration
@@ -87,8 +82,13 @@ internal class Puck2D: Puck {
             ]
 
             var bearing: Double = 0.0
-            if let latestBearing = location.heading {
-                bearing = latestBearing.trueHeading
+            switch puckBearingSource {
+            case .heading:
+                if let latestBearing = location.heading {
+                    bearing = latestBearing.trueHeading
+                }
+            case .course:
+                bearing = location.course
             }
 
             do {