+++ Tests/MapboxMapsTests/MapView/Integration Tests/Map/MapboxMapIntegrationTests.swift
@@ -4,16 +4,24 @@ import MapboxMaps
 class MapboxMapIntegrationTests: IntegrationTestCase {
     var rootView: UIView!
     var mapView: MapView!
+    var dataPathURL: URL!
 
     override func setUpWithError() throws {
+        try guardForMetalDevice()
+
         try super.setUpWithError()
 
-        try guardForMetalDevice()
+        dataPathURL = try temporaryCacheDirectory()
 
         guard let root = rootViewController?.view else {
             throw XCTSkip("No valid UIWindow or root view controller")
         }
+
         rootView = root
+    }
+
+    override func tearDownWithError() throws {
+        try super.tearDownWithError()
     }
 
     // MARK: - Tests
@@ -94,7 +102,8 @@ class MapboxMapIntegrationTests: IntegrationTestCase {
     // MARK: - Helpers
 
     private func setupMapView() {
-        let resourceOptions = ResourceOptions(accessToken: accessToken)
+        let resourceOptions = ResourceOptions(accessToken: accessToken,
+                                              dataPathURL: dataPathURL)
         let mapInitOptions = MapInitOptions(resourceOptions: resourceOptions)
         mapView = MapView(frame: rootView.bounds, mapInitOptions: mapInitOptions)
         rootView.addSubview(mapView)